backend/requirements.txt
fastapi
uvicorn
semgrep
gitpython
httpx
pydantic
python-dotenv
google-generativeai
jinja2
prometheus-fastapi-instrumentator
aiosmtplib
